requirements.txt
# for lib (easai folder):
openai

# for tests (easai_tests folder):
python-dotenv

# for build:
build
twine
